refactor: почистил auth папку и файлы
