--- meetups/src/auth/constants.ts
@@ -1,3 +1,0 @@
-export const jwtConstants = {
-  secret: 'Засунуть потом токен в env',
-};
