+++ meetups/src/auth/dto/payloadDto.ts
@@ -1,0 +1,19 @@
+import { Role } from '@prisma/client';
+import { IsInt, IsNotEmpty, IsString } from 'class-validator';
+
+export class userDto {
+  @IsInt()
+  @IsNotEmpty()
+  id: number;
+
+  @IsString()
+  @IsNotEmpty()
+  username: string;
+
+  @IsString()
+  @IsNotEmpty()
+  role?: Role;
+
+  @IsString()
+  sub?: string;
+}
