+++ meetups/src/auth/guards/jwt/jwt.strategy.ts
@@ -1,8 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { PassportStrategy } from '@nestjs/passport';
 import { ExtractJwt, Strategy } from 'passport-jwt';
-
-import { jwtConstants } from './constants';
+import { userDto } from 'src/auth/dto/payloadDto';
 
 @Injectable()
 export class JwtStrategy extends PassportStrategy(Strategy) {
@@ -10,11 +9,12 @@ export class JwtStrategy extends PassportStrategy(Strategy) {
     super({
       jwtFromRequest: ExtractJwt.fromAuthHeaderAsBearerToken(),
       ignoreExpiration: false,
-      secretOrKey: jwtConstants.secret,
+      secretOrKey: process.env.JWT_ACCESS_SECRET || 'default_access_secret',
     });
   }
 
-  async validate(payload: any) {
+  async validate(payload: userDto) {
+    console.log('JwtStrategy', payload);
     return { id: payload.sub, username: payload.username, role: payload.role };
   }
 }
