+++ meetups/src/auth/guards/local/local.strategy.ts
@@ -2,7 +2,8 @@ import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { PassportStrategy } from '@nestjs/passport';
 import { Strategy } from 'passport-local';
 
-import { AuthService } from './auth.service';
+import { AuthService } from '../../auth.service';
+import { userDto } from 'src/auth/dto/payloadDto';
 
 @Injectable()
 export class LocalStrategy extends PassportStrategy(Strategy) {
@@ -10,8 +11,8 @@ export class LocalStrategy extends PassportStrategy(Strategy) {
     super();
   }
 
-  async validate(username: string, password: string): Promise<any> {
-    //console.log('1) PassportStrategy local validate');
+  async validate(username: string, password: string): Promise<userDto | null> {
+    console.log('1) PassportStrategy local validate');
     const user = await this.authService.validateUser(username, password);
     if (!user) {
       throw new UnauthorizedException();
